Append filter does not correctly handle Polyhedral cells
After clipping a dataset, the appendfilter is run to merge cells
from different inputs, but it breaks some polyhedral cells.
In fact 99% of the time we do not need thre apend filter, so
only execute it if one of the lists actually contains data, otherwise skip it.
This fix saves time and prevents a bug from crashing paraview, but it does not
fix the underlying problem with the append filter.