support new transfer function editing modes in gui
all gui changes are in pqColorOpacityEditorWidget.

added support of 2D, scalar gaussian, gradient gaussian,
and gradient linear transfer functions to fixedpointraycaster
added the necessary helpers to the raycaster.

added limited support of scalar gaussian, gradient gaussian, and gradient l
inear transfer functions to texturemapper3d.
In some cases the spacing may cause problems with getting the
lookuptable from the gradient functions.
the 2D function is NOT supported by the texturemapper.

the vtkPiecewiseFunction now inherits from vtkAbstractPiecewiseFunction.
the gaussian one as well. This allows the volumeproperty to switch between the
two when rendering.
Note - the 2D transfer function does NOT inherit from vtkAbstractPiecewiseFunction.

altered vtkImageGradientMagnitude to be more versatile (handles different
types).

added vtkDiscretizableColorTransferFunctionCollection that works as the new proxy
(instead of  vtkDiscretizableColorTransferFunction).
this includes all the gaussian, linear and 2D functions.

vtkPImageAccumulate added, it is the parallel version of vtkImageAccumulate

added testing for some of the rendering.
WARNING: the png files have to be added to the correct directory
(not included in this commit), or they won't work.

Change-Id: I7a931d2d589e71d78e1952e0b17a517b6147e284