Improve GenerateRepresentation for CellTreeLocator to give tight BBoxes
The CellTreeLocator does not store bounding boxes for each node, but
generates them by bisecting the parent node along a split plane.
The extents of the boxes are therefore always the maximal size for each
node. This patch queries the cells in each node to produce a tight bbox
for each node which improves visualization of the tree.